move modal sources: vm-manage-gpu
