+++ src/integrations/showVmManageGpu.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { renderComponent } from '_/utils/react-modals'
 
-import GpuDataProvider from '_/components/modals/vm/gpu/GpuDataProvider'
-import ManageGpuModal from '_/components/modals/vm/gpu/ManageGpuModal'
+import GpuDataProvider from '_/modals/vm-manage-gpu/GpuDataProvider'
+import ManageGpuModal from '_/modals/vm-manage-gpu/ManageGpuModal'
 
 export function showVmManageGpuModal (vm) {
   renderComponent(
