+++ src/modals/vm-manage-gpu/GpuTableRowDetail.js
