+++ src/modals/vm-manage-gpu/ManageGpuModal.js
