+++ src/modals/vm-manage-gpu/ManageGpuModalBody.js
