+++ src/modals/vm-manage-gpu/handleNonAvailableValue.js
